+++ packages/amplify-category-auth/src/index.js
@@ -7,14 +7,13 @@ const sequential = require('promise-sequential');
 
 const defaults = require('./provider-utils/awscloudformation/assets/cognito-defaults');
 const { getAuthResourceName } = require('./utils/getAuthResourceName');
+const { updateConfigOnEnvInit, migrate } = require('./provider-utils/awscloudformation');
 const {
-  updateConfigOnEnvInit,
   copyCfnTemplate,
   saveResourceParameters,
-  ENV_SPECIFIC_PARAMS,
-  migrate,
   removeDeprecatedProps,
-} = require('./provider-utils/awscloudformation');
+} = require('./provider-utils/awscloudformation/utils/synthesize-resources');
+const { ENV_SPECIFIC_PARAMS } = require('./provider-utils/awscloudformation/constants');
 
 const { transformUserPoolGroupSchema } = require('./provider-utils/awscloudformation/utils/transform-user-pool-group');
 const { uploadFiles } = require('./provider-utils/awscloudformation/utils/trigger-file-uploader');
@@ -45,21 +44,6 @@ async function add(context) {
         return;
       }
       return providerController.addResource(context, result.service);
-    })
-    .then(resourceName => {
-      const options = {
-        service: resultMetadata.service,
-        providerPlugin: resultMetadata.providerName,
-      };
-      const resourceDirPath = path.join(amplify.pathManager.getBackendDirPath(), 'auth', resourceName, 'parameters.json');
-      const authParameters = amplify.readJsonFile(resourceDirPath);
-
-      if (authParameters.dependsOn) {
-        options.dependsOn = authParameters.dependsOn;
-      }
-      amplify.updateamplifyMetaAfterResourceAdd(category, resourceName, options);
-      context.print.success('Successfully added auth resource');
-      return resourceName;
     })
     .catch(err => {
       context.print.info(err.stack);