+++ packages/amplify-category-auth/src/provider-utils/awscloudformation/handlers/get-add-auth-handler.ts
@@ -5,7 +5,8 @@ import { getPostAddAuthMetaUpdater } from '../utils/post-add-auth-meta-update';
 import { getPostAddAuthMessagePrinter } from '../utils/post-add-auth-message-printer';
 
 /**
- * Factory function that returns a ServiceQuestionsResult consumer that handles all of the resource generation logic
+ * Factory function that returns a ServiceQuestionsResult consumer that handles all of the resource generation logic.
+ * The consumer returns the resourceName of the generated resource.
  * @param context The amplify context
  */
 export const getAddAuthHandler = (context: any) => async (request: ServiceQuestionsResult) => {
@@ -14,17 +15,19 @@ export const getAddAuthHandler = (context: any) => async (request: ServiceQuesti
   let projectName = context.amplify.getProjectConfig().projectName.toLowerCase();
   const disallowedChars = /[^A-Za-z0-9]+/g;
   projectName = projectName.replace(disallowedChars, '');
-  return getAddAuthDefaultsApplier(
-    defaultValuesFilename,
-    projectName,
-  )(request)
-    .then(getResourceSynthesizer(context, cfnFilename, provider))
-    .then(request => request.resourceName!)
-    .then(getPostAddAuthMetaUpdater(context, { service: request.serviceName, providerName: provider }))
+  const requestWithDefaults = await getAddAuthDefaultsApplier(defaultValuesFilename, projectName)(request);
+  await getResourceSynthesizer(
+    context,
+    cfnFilename,
+    provider,
+  )(requestWithDefaults)
+    .then(req => req.resourceName!)
+    .then(getPostAddAuthMetaUpdater(context, { service: requestWithDefaults.serviceName, providerName: provider }))
     .then(getPostAddAuthMessagePrinter(context))
     .catch(err => {
       context.print.info(err.stack);
       context.print.error('There was an error adding the auth resource');
       context.usageData.emitError(err);
     });
+  return requestWithDefaults.resourceName!;
 };