+++ packages/amplify-category-function/src/provider-utils/awscloudformation/utils/cloudformationHelpers.ts
@@ -1,9 +1,10 @@
-import { category as categoryName } from '../../../constants';
+import { categoryName, appsyncTableSuffix } from './constants';
+import { getAppSyncResourceName } from './appSyncHelper';
 
 export function getNewCFNEnvVariables(oldCFNEnvVariables, currentDefaults, newCFNEnvVariables, newDefaults) {
   const currentResources = [];
   const newResources = [];
-  const deletedResources = [];
+  let deletedResources = [];
 
   if (currentDefaults.permissions) {
     Object.keys(currentDefaults.permissions).forEach(category => {
@@ -27,8 +28,21 @@ export function getNewCFNEnvVariables(oldCFNEnvVariables, currentDefaults, newCF
     }
   });
 
-  const toBeDeletedEnvVariables = [];
+  const deleteAppSyncTableResources = deletedResources.filter(resource => resource.includes(appsyncTableSuffix.toUpperCase()));
+  deletedResources = deletedResources.filter(resource => !resource.includes(appsyncTableSuffix.toUpperCase()));
+  deleteAppSyncTableResources.forEach(table => {
+    const appsyncResourceName = getAppSyncResourceName();
+    const replacementTableSuffix = `:${appsyncTableSuffix.toUpperCase()}_`;
+    const modelEnvPrefix = `API_${appsyncResourceName.toUpperCase()}_${table
+      .replace(replacementTableSuffix, 'TABLE')
+      .replace('STORAGE_', '')}`;
+    const modelEnvNameKey = `${modelEnvPrefix}_NAME`;
+    const modelEnvArnKey = `${modelEnvPrefix}_ARN`;
+    deletedResources.push(modelEnvNameKey);
+    deletedResources.push(modelEnvArnKey);
+  });
 
+  const toBeDeletedEnvVariables = [];
   Object.keys(oldCFNEnvVariables).forEach(envVar => {
     for (let i = 0; i < deletedResources.length; i += 1) {
       if (envVar.includes(deletedResources[i])) {