+++ packages/amplify-category-analytics/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.17.14](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@2.17.13...amplify-category-analytics@2.17.14) (2020-09-09)
+
+**Note:** Version bump only for package amplify-category-analytics
+
+
+
+
+
 ## [2.17.13](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@2.17.12...amplify-category-analytics@2.17.13) (2020-09-03)
 
 **Note:** Version bump only for package amplify-category-analytics