+++ packages/amplify-category-interactions/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.14](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-interactions@2.4.13...amplify-category-interactions@2.4.14) (2020-09-09)
+
+**Note:** Version bump only for package amplify-category-interactions
+
+
+
+
+
 ## [2.4.13](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-interactions@2.4.12...amplify-category-interactions@2.4.13) (2020-09-03)
 
 **Note:** Version bump only for package amplify-category-interactions