+++ packages/amplify-category-notifications/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.15.23](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-notifications@2.15.22...amplify-category-notifications@2.15.23) (2020-09-09)
+
+**Note:** Version bump only for package amplify-category-notifications
+
+
+
+
+
 ## [2.15.22](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-notifications@2.15.21...amplify-category-notifications@2.15.22) (2020-09-03)
 
 **Note:** Version bump only for package amplify-category-notifications